fixing git push error
