+++ src/main/resources/templates/login.html
@@ -1,8 +1,11 @@
 <!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
 <head>
   <meta charset="UTF-8" />
   <title>Log In</title>
+  <link rel="stylesheet" th:href="@{/css/style.css}">
   <script src="https://kit.fontawesome.com/f692bfbf5c.js" crossorigin="anonymous"></script>
+<<<<<<< Updated upstream
   <style>
     *{
       font-family: Poppins, sans-serif;
@@ -53,29 +56,129 @@
       padding: 0 0 0 45px;
       background: rgba(255,255,255,0.2);
     }
+=======
+<!--  <style th:inline="text">-->
+<!--    /*body{*/-->
+<!--    /*  background: url("@{/img/project.webp}") no-repeat center center fixed;*/-->
+<!--    /*}*/-->
+>>>>>>> Stashed changes
 
-    .input-field {
-      display: flex;
-      flex-direction: column;
-    }
-    label {
-      color:#ffffff;
-    }
+<!--  </style>-->
+<!--  <style>-->
+<!--    *{-->
+<!--      font-family: Poppins, sans-serif;-->
+<!--    }-->
+<!--    body {-->
+<!--      /*background-image:url("/img/project.webp");*/-->
+<!--      /*background-color:black;*/-->
+<!--      background-size: cover;-->
+<!--      background-repeat: no-repeat;-->
+<!--      background-position: center;-->
+<!--      background-attachment: fixed;-->
+<!--    }-->
+<!--    .box {-->
+<!--      display: flex;-->
+<!--      justify-content: center;-->
+<!--      align-items: center;-->
+<!--      min-height: 90vh;-->
+<!--    }-->
+<!--    .container {-->
+<!--      width: 350px;-->
+<!--      display: flex;-->
+<!--      flex-direction: column;-->
+<!--      padding: 0 15px 0 15px;-->
+<!--    }-->
+<!--    span {-->
+<!--      color: #ffffff;-->
+<!--      font-size: small;-->
+<!--      display: flex;-->
+<!--      justify-content: center;-->
+<!--      padding: 10px 0 10px 0;-->
+<!--    }-->
+<!--    header {-->
+<!--      color: #ffffff;-->
+<!--      font-size: 30px;-->
+<!--      display: flex;-->
+<!--      justify-content: center;-->
+<!--      padding: 10px 0 10px 0;-->
+<!--    }-->
+<!--    .input {-->
+<!--      height: 45px;-->
+<!--      width: 87%;-->
+<!--      border: none;-->
+<!--      outline: none;-->
+<!--      border-radius: 30px;-->
+<!--      color: #ffffff;-->
+<!--      padding: 0 0 0 45px;-->
+<!--      background: rgba(255,255,255,0.2);-->
+<!--    }-->
+
+<!--    .input-field {-->
+<!--      display: flex;-->
+<!--      flex-direction: column;-->
+<!--    }-->
+<!--    label {-->
+<!--      color:#ffffff;-->
+<!--    }-->
 
 
-    .material-symbols-outlined {
-      font-variation-settings:
-              'FILL' 0,
-              'wght' 400,
-              'GRAD' 0,
-              'opsz' 48
-    }
+<!--    .material-symbols-outlined {-->
+<!--      font-variation-settings:-->
+<!--              'FILL' 0,-->
+<!--              'wght' 400,-->
+<!--              'GRAD' 0,-->
+<!--              'opsz' 48-->
+<!--    }-->
 
-    i {
-      position: relative;
-      top:-33px;
-      left: 17px;
-    }
+<!--    i {-->
+<!--      position: relative;-->
+<!--      top:-33px;-->
+<!--      left: 17px;-->
+<!--    }-->
+
+<!--    * {-->
+<!--      box-sizing: border-box;-->
+<!--      margin: 0;-->
+<!--      padding: 0;-->
+<!--      /*background-color: #2a2b3c ;*/-->
+<!--    }-->
+
+<!--    nav {-->
+<!--      background-color: #2a2b3c-->
+<!--    }-->
+
+<!--    li,a, button{-->
+<!--      /*font-family: Montserrat, sans-serif;*/-->
+<!--      font-family: "Andale Mono";-->
+<!--      font-weight: 500;-->
+<!--      font-size: 16px;-->
+<!--      color: #037DD6;-->
+<!--      text-decoration: none;-->
+<!--    }-->
+
+<!--    header {-->
+<!--      display: flex;-->
+<!--      justify-content: space-between;-->
+<!--      align-items: center;-->
+<!--      padding: 30px 10px;-->
+<!--    }-->
+<!--    .nav-links {-->
+<!--      list-style: none;-->
+
+<!--    }-->
+
+<!--    .nav-links li {-->
+<!--      display: inline-block;-->
+<!--      padding: 0px 20px;-->
+<!--    }-->
+
+<!--    .nav-links li a {-->
+<!--      transition: all 0.3s ease 0s;-->
+<!--    }-->
+
+<!--    .nav-links li a:hover{-->
+<!--      color: white;-->
+<!--    }-->
 
     li a{
       font-weight: 400;
@@ -102,10 +205,35 @@
     }
 
 
-  </style>
+<!--    button:hover {-->
+<!--      background-color: rgba(0,136,169, 0.8);-->
+<!--    }-->
+
+<!--    /*nav-bar search*/-->
+<!--    .content {-->
+<!--      background: white;-->
+<!--      width: 70vw;-->
+<!--      padding: 3em;-->
+<!--      box-shadow: 0 03em rgba(0,0,0, 0.15);-->
+<!--      position: relative;-->
+<!--    }-->
+
+
+<!--    .search-bar input {-->
+<!--      flex-grow: 1;-->
+<!--      font-size: 1.5rem;-->
+<!--    }-->
+<!--  style="background-image : url('../static/img/project.webp')"-->
+
+<!--  </style>-->
 </head>
 <body>
+<<<<<<< Updated upstream
 <nav th:replace="partials/Nav.html"></nav>
+=======
+<nav th:replace="~{partials/navBar :: navbar}"></nav>
+
+>>>>>>> Stashed changes
 <div th:if="${param.error}">
   Invalid Username and Password.
 </div>
@@ -113,6 +241,7 @@
   You have been logged out.
 </div>
 <div class="box">
+<!--  <nav th:replace="~{partials/navBar :: navbar}"></nav>-->
   <div class="container">
     <div>
       <span></span>
