--- src/main/resources/templates/showPage.html
@@ -1,142 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns="http://www.w3.org/1999/html">
-<head>
-  <meta charset="UTF-8">
-  <title>Button Info</title>
-  <link rel="stylesheet" href="../css/style.css">
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-  <style>
-    *, :after, :before {
-      box-sizing: border-box;
-    }
-  </style>
-</head>
-<body>
-<!--The Header-->
-<div class="container">
-  <div id="trendingBotton" class="text-end d-flex "></div>
-  <nav class="navbar navbar-expand-lg navbar-light bg-light">
-    <div class="container-fluid">
-      <a class="navbar-brand" href="#">CryptEX</a>
-      <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-        <span class="navbar-toggler-icon"></span>
-      </button>
-      <div class="collapse navbar-collapse" id="navbarSupportedContent">
-        <ul class="navbar-nav me-auto mb-2 mb-lg-0">
-          <li class="nav-item">
-            <a class="nav-link active" aria-current="page" href="#">Home</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="#">Link</a>
-          </li>
-          <li class="nav-item dropdown">
-            <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-bs-toggle="dropdown" aria-expanded="false">
-              Dropdown
-            </a>
-            <ul class="dropdown-menu" aria-labelledby="navbarDropdown">
-              <li><a class="dropdown-item" href="#">Action</a></li>
-              <li><a class="dropdown-item" href="#">Another action</a></li>
-              <li><hr class="dropdown-divider"></li>
-              <li><a class="dropdown-item" href="#">Something else here</a></li>
-            </ul>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link disabled">Disabled</a>
-          </li>
-        </ul>
-        <form class="d-flex">
-          <input class="form-control me-2" type="search" placeholder="Search" aria-label="Search">
-          <button class="btn btn-outline-success" type="submit">Search</button>
-        </form>
-      </div>
-    </div>
-  </nav>
-</div>
-<!--Coin basic info-->
-<div class="container infoM">
-  <div class="row">
-    <div class="col-7">
-      <div class=" col-1 border border-dark rounded text-light bg-dark" ><span>Rank 1</span></div>
-      <div class="fs-4 fw-bold">&#x1F601<span>BitCoin</span></div>
-      <div class=" fs-2 fw-boldcoin-price col-12">10,000</div>
-      <div class=" col-12">0.06222048</div>
-      <!--        button-->
-      <div class="d-flex">
-        <!--          Add Price Alert button-->
-        <div class="border border-dark rounded-3 me-1">&x1F514</div>
-        <!--          Add to Portfolio and track coin price-->
-        <div class="border border-dark rounded-3 me-1">&#x2605</div>
-        <!--          watchlist-->
-        <div class="border border-dark rounded-3 cal-2 smallbt">On 1,151,579 watchlist</div>
-      </div>
-      <!--			pill range-->
-      <div class=" col-5 boxLoader rounded-pill mt-3"></div>
-      <div class="col-5 boxLoader1 mt-3 d-flex justify-content-center">
-        <div class="gridText">$29,000</div>
-        <div class="margPush gridText">24H Rang</div>
-        <div class="gridText">$30,000</div>
-      </div>
-      <!--			Market UL -->
-      <div class=" col-12 d-flex">
-        <div class="col-4">
-          <div class="d-flex justify-content-between col-3">
-            <div class="d-flex justify-content-start">
-              <span class="coin-marketcap col-12 gridText">${marketCap1}</span>
-              <span class="coin-marketcap col-12 gridText" style="margin-left: 40px">$1,000</span>
-            </div>
-          </div>
-          <div class="d-flex justify-content-between col-3">
-            <div class="d-flex justify-content-start">
-              <span class="coin-marketcap col-12 gridText">${marketCap1}</span>
-              <span class="coin-marketcap col-12 gridText" style="margin-left: 40px">$1,000</span>
-            </div>
-          </div>
-          <div class="d-flex justify-content-between col-3">
-            <div class="d-flex justify-content-start">
-              <span class="coin-marketcap col-12 gridText">${marketCap1}</span>
-              <span class="coin-marketcap col-12 gridText" style="margin-left: 40px">$1,000</span>
-            </div>
-          </div>
-        </div>
-        <div class="col-4">
-          <div class="d-flex justify-content-between col-3">
-            <div class="d-flex justify-content-start">
-              <span class="coin-marketcap col-12 gridText">${marketCap1}</span>
-              <span class="coin-marketcap col-12 gridText" style="margin-left: 40px">$1,000</span>
-            </div>
-          </div>
-          <div class="d-flex justify-content-between col-3">
-            <div class="d-flex justify-content-start">
-              <span class="coin-marketcap col-12 gridText">${marketCap1}</span>
-              <span class="coin-marketcap col-12 gridText" style="margin-left: 40px">$1,000</span>
-            </div>
-          </div>
-          <div class="d-flex justify-content-between col-3">
-            <div class="d-flex justify-content-start">
-              <span class="coin-marketcap col-12 gridText">${marketCap1}</span>
-              <span class="coin-marketcap col-12 gridText" style="margin-left: 40px">$1,000</span>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class=" Borders col-4 border border-secondary">
-      <h1 class="fs-4 fw-bold">Info</h1>
-      <div class="d-flex justify-content-start">
-        <span class="coin-marketcap col-12 gridText">Website</span>
-        <span class="coin-marketcap border border-dark rounded text-dark gridText"style="margin-right: 400px">bitcoin.org</span>
-      </div>
-      <div class="d-flex justify-content-start">
-        <span class="coin-marketcap col-12 gridText">Community</span>
-        <span class="coin-marketcap border border-dark rounded text-dark gridText"style="margin-right: 400px">Twitter</span>
-        <span class="coin-marketcap border border-dark rounded text-dark gridText"style="margin-right: 400px">Facebook</span>
-      </div>
-    </div>
-    <!--		<div class=" Borders col-4 bg-primary border border-secondary"></div>-->
-  </div>
-</div>
-
-<script src="https://code.jquery.com/jquery-3.6.4.min.js" integrity="sha256-oP6HI9z1XaZNBrJURtCoUT5SUnxFr8s3BzRl+cbzUq8=" crossorigin="anonymous"></script>
-<script src="../js/showButton.js"></script>
-</body>
-</html>
